Move ApolloProvider to App.js
